Fix "LATIN SMALL LETTER I" to "CYRILLIC SMALL LETTER BYELORUSSIAN-UKRAINIAN I".
It is possible that fonts have different glyph for this letter.

I think that original translation was prepared in build-in editor of popular
"Far" file manager which is console application and it usually run in cp866
codepage which doesn't have codepoint for "CYRILLIC SMALL UKRAINIAN I". (xD)

But most valuable reason to make this change is to have an ability to run
ispell like program.

I also fix some spelling errors, so ispell with 'ukrainian' dictionary doesn't
show any errors by now.


git-svn-id: https://nsis.svn.sourceforge.net/svnroot/nsis/NSIS/trunk@6286 212acab6-be3b-0410-9dea-997c60f758d6